Add a working example for data-querys on the jena-db
